# Conflicts:
#	package-lock.json

+++ shared/src/business/useCases/docketEntry/updateDocketEntryInteractor.js
@@ -1,4 +1,7 @@
 const {
+  aggregatePartiesForService,
+} = require('../../utilities/aggregatePartiesForService');
+const {
   isAuthorized,
   ROLE_PERMISSIONS,
 } = require('../../../authorization/authorizationClientService');
@@ -20,6 +23,7 @@ const { UnauthorizedError } = require('../../../errors/errors');
 exports.updateDocketEntryInteractor = async ({
   applicationContext,
   documentMetadata,
+  isSavingForLater,
   primaryDocumentFileId,
 }) => {
   const authorizedUser = applicationContext.getCurrentUser();
@@ -111,11 +115,12 @@ exports.updateDocketEntryInteractor = async ({
     const workItemToDelete = currentDocument.workItems.find(
       workItem => !workItem.document.isFileAttached,
     );
-
-    await applicationContext.getPersistenceGateway().deleteWorkItemFromInbox({
-      applicationContext,
-      workItem: workItemToDelete,
-    });
+    if (workItemToDelete) {
+      await applicationContext.getPersistenceGateway().deleteWorkItemFromInbox({
+        applicationContext,
+        workItem: workItemToDelete,
+      });
+    }
 
     const workItem = documentEntity.getQCWorkItem();
     Object.assign(workItem, {
@@ -150,6 +155,21 @@ exports.updateDocketEntryInteractor = async ({
     });
 
     documentEntity.addWorkItem(workItem);
+
+    if (!isSavingForLater) {
+      const servedParties = aggregatePartiesForService(caseEntity);
+
+      documentEntity.setAsServed(servedParties.all);
+    } else {
+      documentEntity.numberOfPages = await applicationContext
+        .getUseCaseHelpers()
+        .countPagesInDocument({
+          applicationContext,
+          documentId: primaryDocumentFileId,
+        });
+
+      caseEntity.updateDocument(documentEntity);
+    }
 
     await applicationContext
       .getPersistenceGateway()
